fix: remove dead Runtime worker-pool path after dispatch cutover
Delete the superseded Runtime-owned pool and workerPoolDispatchResultHook implementation that live construction no longer wires after the Workers boundary cutover. Refresh deadcode and unit-coverage baselines for the live package surface.
