+++ docs/internal/baselines/deadcode-baseline.txt
@@ -110,10 +110,6 @@ pkg/services/factory_runtime/engine/options_test.go: unreachable func: WithResul
 pkg/services/factory_runtime/engine/options_test.go: unreachable func: WithResultBufferDrainObserver
 pkg/services/factory_runtime/engine/options_test.go: unreachable func: WithTokenTransformer
 pkg/services/factory_runtime/engine/options_test.go: unreachable func: WithWorkstationResponseRecorder
-pkg/services/factory_runtime/runtime/dispatch_result_hook.go: unreachable func: workerPoolDispatchResultHook.HasBufferedResults
-pkg/services/factory_runtime/runtime/dispatch_result_hook.go: unreachable func: workerPoolDispatchResultHook.HasPendingResults
-pkg/services/factory_runtime/runtime/dispatch_result_hook.go: unreachable func: workerPoolDispatchResultHook.SetOnBufferedResult
-pkg/services/factory_runtime/runtime/dispatch_result_hook.go: unreachable func: workerPoolDispatchResultHook.SignalBufferedResults
 pkg/services/factory_runtime/runtime/factory.go: unreachable func: WorkflowContext
 pkg/services/factory_runtime/runtime/factory_event_history_test.go: unreachable func: assertBatchRequestReplayEvents
 pkg/services/factory_runtime/runtime/factory_event_history_test.go: unreachable func: assertDispatchResponseCount
@@ -142,6 +138,7 @@ pkg/services/factory_runtime/runtime/factory_runtime_test_helpers_test.go: unrea
 pkg/services/factory_runtime/runtime/factory_runtime_test_helpers_test.go: unreachable func: safeBoundaryUnsafeValues
 pkg/services/factory_runtime/runtime/factory_runtime_test_helpers_test.go: unreachable func: sliceValueForRuntimeTest
 pkg/services/factory_runtime/runtime/factory_runtime_test_helpers_test.go: unreachable func: stringValueForRuntimeTest
+pkg/services/factory_runtime/runtime/root_contract_service_test.go: unreachable func: testRuntimeClock.Now
 pkg/services/factory_runtime/service/config.go: unreachable func: ValidateRecordReplayPaths
 pkg/services/factory_runtime/service/host/bundle.go: unreachable func: FormatRuntimeStartTime
 pkg/services/factory_runtime/service/host/lifecycle.go: unreachable func: IsCanceledStartup
